web/backend/requirements/production.txt
# Production requirements
-r base.txt

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
sentry-sdk==1.38.0
django-health-check==3.17.0

# Performance
django-cachalot==2.6.1
django-redis==5.4.0

# Security
django-csp==3.7
django-permissions-policy==4.18.0
